requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
python-telegram-bot==20.7
reportlab==4.0.7
openpyxl==3.1.2
python-dotenv==1.0.0
Flask-WTF==1.2.1
email-validator==2.1.0
Flask-Mail==0.9.1
APScheduler==3.10.4
qrcode[pil]==7.4.2
requests==2.31.0
Pillow==10.4.0
Flask-Babel==4.0.0